+++ source/gamecart/gamecart.h
@@ -23,5 +23,5 @@ typedef struct {
 u32 GetCartName(char* name, CartData* cdata);
 u32 InitCardRead(CartData* cdata);
 u32 ReadCartSectors(u8* buffer, u32 sector, u32 count, CartData* cdata);
-u32 ReadCartBytes(u8* buffer, u32 offset, u32 count, CartData* cdata);
-u32 ReadCartPrivateHeader(u8* buffer, u32 offset, u32 count, CartData* cdata);
+u32 ReadCartBytes(u8* buffer, u64 offset, u64 count, CartData* cdata);
+u32 ReadCartPrivateHeader(u8* buffer, u64 offset, u64 count, CartData* cdata);
