+++ source/virtual/vmem.h
@@ -4,5 +4,5 @@
 #include "virtual.h"
 
 bool ReadVMemDir(VirtualFile* vfile, VirtualDir* vdir);
-int ReadVMemFile(const VirtualFile* vfile, u8* buffer, u32 offset, u32 count);
-int WriteVMemFile(const VirtualFile* vfile, const u8* buffer, u32 offset, u32 count);
+int ReadVMemFile(const VirtualFile* vfile, u8* buffer, u64 offset, u64 count);
+int WriteVMemFile(const VirtualFile* vfile, const u8* buffer, u64 offset, u64 count);
